+++ R/randomizer.R
@@ -10,6 +10,8 @@
 #' @param label Name to use for the generated condition variable
 #' @param seed Random seed to use as the starting point
 #' @param data_filter Expression entered in, e.g., Data > View to filter the dataset in Radiant. The expression should be a string (e.g., "price > 10000")
+#' @param arr Expression to arrange (sort) the data on (e.g., "color, desc(price)")
+#' @param rows Rows to select from the specified dataset
 #' @param na.rm Remove rows with missing values (FALSE or TRUE)
 #' @param envir Environment to extract data from
 #'
@@ -28,7 +30,10 @@ randomizer <- function(dataset, vars,
                        conditions = c("A", "B"),
                        blocks = NULL, probs = NULL,
                        label = ".conditions",
-                       seed = 1234, data_filter = "",
+                       seed = 1234,
+                       data_filter = "",
+                       arr = "",
+                       rows = NULL,
                        na.rm = FALSE,
                        envir = parent.frame()) {
   df_name <- if (is_string(dataset)) dataset else deparse(substitute(dataset))
@@ -37,7 +42,7 @@ randomizer <- function(dataset, vars,
     vars <- c(vars, blocks)
   }
 
-  dataset <- get_data(dataset, vars, filt = data_filter, na.rm = na.rm, envir = envir)
+  dataset <- get_data(dataset, vars, filt = data_filter, arr = arr, rows = rows, na.rm = na.rm, envir = envir)
 
   ## use seed if provided
   seed <- gsub("[^0-9]", "", seed)
@@ -97,6 +102,12 @@ summary.randomizer <- function(object, dec = 3, ...) {
   cat("Data         :", object$df_name, "\n")
   if (!is.empty(object$data_filter)) {
     cat("Filter       :", gsub("\\n", "", object$data_filter), "\n")
+  }
+  if (!is.empty(object$arr)) {
+    cat("Arrange      :", gsub("\\n", "", object$arr), "\n")
+  }
+  if (!is.empty(object$rows)) {
+    cat("Slice        :", gsub("\\n", "", object$rows), "\n")
   }
   if (!is.empty(object$blocks)) {
     cat("Variables    :", setdiff(object$vars, object$blocks), "\n")