+++ inst/app/tools/analysis/randomizer_ui.R
@@ -212,7 +212,7 @@ observeEvent(input$rndr_store, {
   register(dataset)
   updateSelectInput(session, "dataset", selected = input$dataset)
 
-  ## See https://shiny.rstudio.com/reference/shiny/latest/modalDialog.html
+  ## See https://shiny.posit.co/reference/shiny/latest/modalDialog.html
   showModal(
     modalDialog(
       title = "Data Stored",